feat: add delete data feature
